adding scopes data file
